% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{duck_exec}
\alias{duck_exec}
\title{Execute a statement for the default connection}
\usage{
duck_exec(sql, ..., con = NULL)
}
\arguments{
\item{sql}{The statement to run.}

\item{...}{These dots are for future extensions and must be empty.}

\item{con}{The connection, defaults to the default connection.}
}
\value{
The return value of the \code{\link[DBI:dbExecute]{DBI::dbExecute()}} call, invisibly.
}
\description{
The \pkg{duckplyr} package relies on a DBI connection
to an in-memory database.
The \code{duck_exec()} function allows running SQL statements
with this connection to, e.g., set up credentials
or attach other databases.
See \url{https://duckdb.org/docs/configuration/overview.html}
for more information on the configuration options.
}
\examples{
duck_exec("SET threads TO 2")
}
\seealso{
\code{\link[=duck_sql]{duck_sql()}}
}
