% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{duckdb_from_file}
\alias{duckdb_from_file}
\title{TODO}
\usage{
duckdb_from_file(path, table_function, options = list())
}
\arguments{
\item{path}{Path to file TODO add what kind}

\item{table_function}{TODO}

\item{options}{TODO}
}
\description{
TODO
}
