% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckplyr-package.R
\docType{package}
\name{duckplyr-package}
\alias{duckplyr}
\alias{duckplyr-package}
\title{duckplyr: A DuckDB-backed version of dplyr}
\description{
A drop-in replacement for dplyr, powered by DuckDB for performance. Also defines a set of generics that provide a low-level implementer's interface for the high-level user interface of dplyr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/duckdblabs/duckplyr}
  \item Report bugs at \url{https://github.com/duckdblabs/duckplyr/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Hannes Mühleisen (\href{https://orcid.org/0000-0001-8552-0029}{ORCID})
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
