% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-rd.R, R/explain.R
\name{explain.duckplyr_df}
\alias{explain.duckplyr_df}
\title{Explain details of a tbl}
\usage{
\method{explain}{duckplyr_df}(x, ...)
}
\arguments{
\item{x}{An object to explain}

\item{...}{Other parameters possibly used by generic}
}
\value{
The input, invisibly.
}
\description{
This is a method for the \code{\link[dplyr:explain]{dplyr::explain()}} generic.
This is a generic function which gives more details about an object than \code{print()},
and is more focused on human readable output than \code{str()}.
}
\examples{
library(duckplyr)
df <- duckdb_tibble(x = c(1, 2))
df <- mutate(df, y = 2)
explain(df)
}
\seealso{
\code{\link[dplyr:explain]{dplyr::explain()}}
}
