% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{expr}
\alias{expr}
\alias{new_expr}
\alias{expr_reference}
\alias{expr_constant}
\alias{expr_function}
\alias{expr_set_alias}
\title{Relational expressions}
\usage{
new_expr(x, class = NULL)

expr_reference(name, rel = NULL, alias = NULL)

expr_constant(val, alias = NULL)

expr_function(name, args, alias = NULL)

expr_set_alias(expr, alias = NULL)
}
\arguments{
\item{x}{An object.}

\item{class}{Classes added in front of the `"relational_expr"` base class.}

\item{name}{The name of the column or function to reference.}

\item{rel}{The name of the relation to reference.}

\item{alias}{An alias for the new expression.}

\item{val}{The value to use in the constant expression.}

\item{args}{Function arguments, a list of `expr` objects.}

\item{expr}{An `expr` object.}
}
\description{
TBD.
}
