% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{ext_install}
\alias{ext_install}
\title{Install duckdb extension}
\usage{
ext_install(force = FALSE)
}
\arguments{
\item{force}{Force installation.}
}
\description{
This package requires a duckdb extension to be installed.
These function give control over the installation process.
The installation is tied to the R installation.
}
\examples{
if (FALSE) {
  ext_install(force = TRUE)
}
}
