% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\name{flights_df}
\alias{flights_df}
\title{Flight data}
\usage{
flights_df()
}
\description{
Provides a copy of \code{nycflights13::flights} that is compatible with duckplyr,
as a tibble.
Call \code{\link[=as_duck_tbl]{as_duck_tbl()}} to enable duckplyr operations.
}
\examples{
\dontshow{if (requireNamespace("nycflights13", quietly = TRUE)) withAutoprint(\{ # examplesIf}
flights_df()
\dontshow{\}) # examplesIf}
}
