% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-rd.R, R/head.R
\name{head.duckplyr_df}
\alias{head.duckplyr_df}
\title{Return the First Parts of an Object}
\usage{
\method{head}{duckplyr_df}(x, n = 6L, ...)
}
\arguments{
\item{x}{A data.frame}

\item{n}{A positive integer, how many rows to return.}

\item{...}{Not used yet.}
}
\description{
This is a method for the \code{\link[=head]{head()}} generic.
See "Fallbacks" section for differences in implementation.
Return the first rows of a data.frame
}
\section{Fallbacks}{

You cannot use \code{head.duckplyr_df}
\itemize{
\item with a negative n.
}

If you do the code will fall back to \code{head()} without any error.
}

\examples{
head(mtcars, 2)
}
\seealso{
\code{\link[=head]{head()}}
}
