% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_duckplyr_df.R
\name{is_duckplyr_df}
\alias{is_duckplyr_df}
\title{Class predicate for duckplyr data frames}
\usage{
is_duckplyr_df(.data)
}
\arguments{
\item{.data}{Data}
}
\description{
Class predicate for duckplyr data frames
}
