% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{last_rel}
\alias{last_rel}
\title{Retrieve details about the most recent computation}
\usage{
last_rel()
}
\value{
A duckdb "relation" object, or \code{NULL} if no computation has been performed yet.
}
\description{
Before a result is computed, it is specified as a "relation" object.
This function retrieves this object for the last computation that led to the materialization of a data frame.
}
