% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckplyr-package.R, R/mutate.R
\name{mutate}
\alias{mutate}
\alias{mutate.duckplyr_df}
\title{Create, modify, and delete columns (duckplyr)}
\usage{
mutate(.data, ...)

\method{mutate}{duckplyr_df}(
  .data,
  ...,
  .by = NULL,
  .keep = c("all", "used", "unused", "none"),
  .before = NULL,
  .after = NULL
)
}
\description{
blabla
}
\examples{
library("duckplyr")
df <- data.frame(x = c(1, 2), row.names = c("a", "b"))
df <- mutate(df, y = 2)
df
}
\seealso{
\code{\link[dplyr:mutate]{dplyr::mutate()}}
}
