% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-supported.R
\name{not-supported}
\alias{not-supported}
\title{Verbs not implemented in duckplyr}
\description{
The following dplyr generics have no counterpart method in duckplyr.
If you want to help add a new verb,
please refer to our contributing guide \url{https://duckplyr.tidyverse.org/CONTRIBUTING.html#support-new-verbs}
}
\section{Unsupported verbs}{

For these verbs, duckplyr will fall back to dplyr.
\itemize{
\item \code{\link[=add_count]{add_count()}}
\item \code{\link[=compute]{compute()}}
\item \code{\link[=cross_join]{cross_join()}}
\item \code{\link[dplyr:do]{dplyr::do()}}
\item \code{\link[dplyr:group_by]{dplyr::group_by()}}
\item \code{\link[dplyr:group_data]{dplyr::group_indices()}}
\item \code{\link[dplyr:group_data]{dplyr::group_keys()}}
\item \code{\link[dplyr:group_map]{dplyr::group_map()}}
\item \code{\link[dplyr:group_map]{dplyr::group_modify()}}
\item \code{\link[dplyr:group_nest]{dplyr::group_nest()}}
\item \code{\link[dplyr:group_data]{dplyr::group_size()}}
\item \code{\link[dplyr:group_split]{dplyr::group_split()}}
\item \code{\link[dplyr:group_trim]{dplyr::group_trim()}}
\item \code{\link[dplyr:group_data]{dplyr::groups()}}
\item \code{\link[dplyr:group_data]{dplyr::n_groups()}}
\item \code{\link[=nest_by]{nest_by()}}
\item \code{\link[=nest_join]{nest_join()}}
\item \code{\link[=reframe]{reframe()}}
\item \code{\link[=rename_with]{rename_with()}}
\item \code{\link[=rows_append]{rows_append()}}
\item \code{\link[=rows_delete]{rows_delete()}}
\item \code{\link[=rows_insert]{rows_insert()}}
\item \code{\link[=rows_patch]{rows_patch()}}
\item \code{\link[=rows_update]{rows_update()}}
\item \code{\link[=rows_upsert]{rows_upsert()}}
\item \code{\link[dplyr:rowwise]{dplyr::rowwise()}}
\item \code{\link[=setequal]{setequal()}}
\item \code{\link[=slice_head]{slice_head()}}
\item \code{\link[=slice_sample]{slice_sample()}}
\item \code{\link[=slice_tail]{slice_tail()}}
\item \code{\link[=slice]{slice()}}
\item \code{\link[=ungroup]{ungroup()}}
}
}

