% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io2-.R
\name{read_duckplyr}
\alias{read_duckplyr}
\title{Read Parquet, CSV, and other files using DuckDB}
\usage{
read_duckplyr(
  path,
  table_function,
  ...,
  lazy = TRUE,
  options = list(),
  class = NULL
)
}
\arguments{
\item{path}{Path to files, glob patterns \code{*} and \verb{?} are supported.}

\item{table_function}{The name of a table-valued
DuckDB function such as \code{"read_parquet"},
\code{"read_csv"}, \code{"read_csv_auto"} or \code{"read_json"}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{lazy}{If \code{TRUE} (the default), \code{\link[=collect]{collect()}} must be called
before the data can be accessed.}

\item{options}{Arguments to the DuckDB function
indicated by \code{table_function}.}

\item{class}{The class of the output.
By default, a tibble is created.
The returned object will always be a data frame.
Use \code{class = "data.frame"} or \code{class = character()}
to create a plain data frame.}
}
\value{
A data frame for \code{df_from_file()}, or a \code{duckplyr_df} for
\code{duckplyr_df_from_file()}, extended by the provided \code{class}.
}
\description{
\code{df_from_file()} uses arbitrary table functions to read data.
See \url{https://duckdb.org/docs/data/overview} for a documentation
of the available functions and their options.
To read multiple files with the same schema,
pass a wildcard or a character vector to the \code{path} argument,
}
