% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_tostring}
\alias{rel_tostring}
\alias{rel_names}
\title{TBD}
\usage{
rel_tostring(rel, ...)

rel_names(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A (new/modified) relational object.

A (new/modified) relational object.
}
\description{
TBD.
}
