% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_aggregate}
\alias{rel_aggregate}
\title{Lazily aggregate a relation object}
\usage{
rel_aggregate(rel, groups, aggregates, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{groups}{a list of DuckDB expressions to group by}

\item{aggregates}{a (optionally named) list of DuckDB expressions with aggregates to compute}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
the now aggregated relation object
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
aggrs <- list(avg_hp = relexpr_function("avg", list(relexpr_reference("hp"))))
rel2 <- rel_aggregate(rel, list(relexpr_reference("cyl")), aggrs)
}
