% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_alias}
\alias{rel_alias}
\title{Get the internal alias for a relation object}
\usage{
rel_alias(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
An alias (character).
}
\description{
TBD.
}
\examples{
mtcars_rel <- new_relational(mtcars, class = "dfrel")
rel_alias.dfrel <- function(rel, ...) tracemem(rel)
rel_alias(mtcars_rel)
}
