% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_explain}
\alias{rel_explain}
\title{Print the EXPLAIN output for a relational object}
\usage{
rel_explain(rel, ...)
}
\arguments{
\item{rel}{The relational object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A (new/modified) relational object.
}
\description{
TBD.
}
\examples{
mtcars_rel <- new_relational(mtcars, class = "dfrel")
rel_explain.dfrel <- function(rel, ...) {
  cat("A relational object")
  print(rel)
  invisible(rel)
}
rel_explain(mtcars_rel)
}
