% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_filter}
\alias{rel_filter}
\title{rel_filter}
\usage{
rel_filter(rel, exprs, ...)
}
\arguments{
\item{rel}{The relational object.}

\item{exprs}{a list of expressions to filter by}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
the now filtered relational object
}
\description{
`rel_filter()` keeps rows that match a predicate, to be used by `dplyr::filter()`.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rel <- rel_from_df(mtcars)
rel2 <- rel_filter(
  rel,
  list(
    relexpr_function(
      "gt",
      list(relexpr_reference("cyl"), relexpr_constant("6")))
  )
)
\dontshow{\}) # examplesIf}
}
