% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_inner_join}
\alias{rel_inner_join}
\title{Lazily INNER join two relation objects}
\usage{
rel_inner_join(left, right, conds, ...)
}
\arguments{
\item{left}{the left-hand-side relation object}

\item{right}{the right-hand-side relation object}

\item{conds}{a list of DuckDB expressions to use for the join}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object resulting from the join
}
\description{
TBD.
}
\examples{
left <- rel_from_df(mtcars)
right <- rel_from_df(mtcars)
cond <- list(expr_function("eq", list(expr_reference("cyl", left), expr_reference("cyl", right))))
rel2 <- rel_inner_join(left, right, cond)
}
