% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_join}
\alias{rel_join}
\title{Lazily INNER join two relation objects}
\usage{
rel_join(
  left,
  right,
  conds,
  join = c("inner", "left", "right", "outer", "cross", "semi", "anti"),
  join_ref_type = c("regular", "natural", "cross", "positional", "asof"),
  ...
)
}
\arguments{
\item{left}{the left-hand-side relation object}

\item{right}{the right-hand-side relation object}

\item{conds}{a list of DuckDB expressions to use for the join}

\item{join}{type of join}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object resulting from the join
}
\description{
TBD.
}
