% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_limit}
\alias{rel_limit}
\title{Lazily limit the rows in a relational object}
\usage{
rel_limit(rel, n, ...)
}
\arguments{
\item{rel}{The relational object.}

\item{n}{The number of rows.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A (new/modified) relational object.
}
\description{
TBD.
}
\examples{
mtcars_rel <- new_relational(list(mtcars), class = "dfrel")
rel_limit.dfrel <- function(rel, n, ...) {
  new_relational(
    head(unclass(rel)[[1]], n),
    class = "dfrel"
  )
}

}
