% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_project}
\alias{rel_project}
\title{Lazily project a relation object}
\usage{
rel_project(rel, exprs, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{exprs}{a list of DuckDB expressions to project}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
the now projected relation object
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
rel2 <- rel_project(rel, list(expr_reference("cyl"), expr_reference("disp")))
}
