% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_set_alias}
\alias{rel_set_alias}
\title{Set the internal alias for a relation object}
\usage{
rel_set_alias(rel, alias, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{alias}{the new alias}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A (new/modified) relational object.
}
\description{
TBD.
}
\examples{
mtcars_rel <- new_relational(mtcars, class = "dfrel")
rel_set_alias.dfrel <- function(rel, alias,...) {
  attr(rel, "alias") <- alias
  rel
}
mtcars_rel <- rel_set_alias(mtcars_rel, "blop")
attr(mtcars_rel, "alias")
}
