% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_set_intersect}
\alias{rel_set_intersect}
\title{Lazily compute a set_intersect result on a relation object}
\usage{
rel_set_intersect(rel_a, rel_b, ...)
}
\arguments{
\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object with the result
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
rel2 <- rel_set_intersect(rel)
}
