% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_set_symdiff}
\alias{rel_set_symdiff}
\title{Lazily compute a set_symdiff result on a relation object}
\usage{
rel_set_symdiff(rel_a, rel_b, ...)
}
\arguments{
\item{rel_a}{a DuckDB relation object}

\item{rel_b}{a DuckDB relation object}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object with the result
}
\description{
TBD.
}
\examples{
rel_a <- new_relational(c(1, 1, 2), class = "vecrel")
rel_b <- new_relational(c(1, 3, 2), class = "vecrel")
rel_set_symdiff.vecrel <- function(rel_a, rel_b, ...) {
class(rel_a) <- setdiff(class(rel_a), "relational")
class(rel_b) <- setdiff(class(rel_b), "relational")
new_relational(unique(c(setdiff(rel_a, rel_b), setdiff(rel_b, rel_a))))
}
rel_set_symdiff(rel_a, rel_b)
}
