% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_to_df}
\alias{rel_to_df}
\title{Convert a relation object to a data frame}
\usage{
rel_to_df(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A data frame.
}
\description{
TBD.
}
\examples{
mtcars_rel <- new_relational(list(mtcars), class = "dfrel")
rel_to_df.dfrel <- function(rel, ...) {
  unclass(rel)[[1]]
}
rel_to_df(mtcars_rel)
}
