% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff-rd.R, R/setdiff.R
\name{setdiff.duckplyr_df}
\alias{setdiff.duckplyr_df}
\title{Set diff}
\usage{
\method{setdiff}{duckplyr_df}(x, y, ...)
}
\arguments{
\item{x, y}{Pair of compatible data frames. A pair of data frames is
compatible if they have the same column names (possibly in different
orders) and compatible types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
This is a method for the \code{\link[dplyr:setops]{dplyr::setdiff()}} generic.
See "Fallbacks" section for differences in implementation.
\code{setdiff(x, y)} finds all rows in \code{x} that aren't in \code{y}.
}
\section{Fallbacks}{

You cannot use \code{setdiff.duckplyr_df}
\itemize{
\item with duplicate names,
\item with tables of different widths,
\item with tables that have different names.
}

If you do the code will fall back to \code{dplyr::setdiff()} without any error.
}

\examples{
df1 <- tibble(x = 1:3)
df2 <- tibble(x = 3:5)
setdiff(df1, df2)
setdiff(df2, df1)
}
\seealso{
\code{\link[dplyr:setops]{dplyr::setdiff()}}
}
