% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats_show}
\alias{stats_show}
\title{Show stats}
\usage{
stats_show()
}
\value{
Called for its side effect.
}
\description{
Prints statistics on how many calls were handled by DuckDB.
The output shows the total number of requests in the current session,
split by fallbacks to dplyr and requests handled by duckdb.
}
\examples{
stats_show()

tibble(a = 1:3) |>
  as_duckplyr_tibble() |>
  mutate(b = a + 1)

stats_show()
}
