% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symdiff-rd.R, R/symdiff.R
\name{symdiff.duckplyr_df}
\alias{symdiff.duckplyr_df}
\title{Symmetric difference}
\usage{
\method{symdiff}{duckplyr_df}(x, y, ...)
}
\arguments{
\item{x, y}{Pair of compatible data frames. A pair of data frames is
compatible if they have the same column names (possibly in different
orders) and compatible types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
This is a method for the \code{\link[dplyr:setops]{dplyr::symdiff()}} generic.
See "Fallbacks" section for differences in implementation.
\code{symdiff(x, y)}  computes the symmetric difference,
i.e. all rows in \code{x} that aren't in \code{y} and all rows in \code{y} that aren't in \code{x}.
}
\section{Fallbacks}{

You cannot use \code{symdiff.duckplyr_df}
\itemize{
\item if column names are duplicated in one of the tables,
\item if column names are different in both tables.
}

If you do the code will fall back to \code{dplyr::symdiff()} without any error.
}

\examples{
df1 <- tibble(x = 1:3)
df2 <- tibble(x = 3:5)
symdiff(df1, df2)
}
\seealso{
\code{\link[dplyr:setops]{dplyr::symdiff()}}
}
