% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_table_to_duckplyr_df.R
\name{table_to_duckplyr_df}
\alias{table_to_duckplyr_df}
\title{create a duckplyr df from a table in a duckdb connection}
\usage{
table_to_duckplyr_df(con, table_name)
}
\arguments{
\item{con}{a connection to a duckdb database}

\item{table_name}{the name of the table in the duckdb database that will be referenced as a duckplyr df}
}
\value{
An object of class \code{"duckplyr_df"}
}
\description{
The con object \strong{MUST} be the default con for duckplyr
The DUCKPLYR_META_SKIP setting will be off for this to work
}
