% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_duckplyr_df.R
\name{as_duckplyr_df}
\alias{as_duckplyr_df}
\title{Convert to a duckplyr data frame}
\usage{
as_duckplyr_df(.data)
}
\arguments{
\item{.data}{data frame or tibble to transform}
}
\description{
Convert to a duckplyr data frame
}
