% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-rd.R, R/collect.R
\name{collect.duckplyr_df}
\alias{collect.duckplyr_df}
\title{Force conversion to a data frame}
\usage{
\method{collect}{duckplyr_df}(x, ...)
}
\arguments{
\item{x}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for more
details.}

\item{...}{Arguments passed on to methods}
}
\description{
This is a method for the \code{\link[dplyr:compute]{dplyr::collect()}} generic.
\code{collect()} converts the input to a tibble, materializing any lazy operations.
}
\details{
Due to the way duckplyr uses ALTREP (Alternative Representation for R),
string columns are not fully materialized in R memory immediately upon calling \code{collect()}.
Instead, memory allocation for strings is deferred until the values are explicitly accessed
(for instance, when printing to the console or calling \code{str()}).
This makes \code{collect()} fast, but doing operations that access all string values
directly after \code{collect()} will suffer a one-time performance hit as the strings are materialized.
}
\examples{
library(duckplyr)
df <- duckdb_tibble(x = c(1, 2), .lazy = TRUE)
df
try(print(df$x))
df <- collect(df)
df
}
\seealso{
\code{\link[dplyr:compute]{dplyr::collect()}}
}
