% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_file.R
\name{compute_file}
\alias{compute_file}
\alias{compute_parquet}
\alias{compute_csv}
\title{Compute results to a file}
\usage{
compute_parquet(x, path, ..., funnel = NULL, options = NULL)

compute_csv(x, path, ..., funnel = NULL, options = NULL)
}
\arguments{
\item{x}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for more
details.}

\item{path}{The path to store the result in.}

\item{...}{These dots are for future extensions and must be empty.}

\item{funnel}{Either a string:
\itemize{
\item \code{"closed"}:  a funneled data frame,
\item \code{"open"}: an unfunneled data frame,
\item \code{"drip"}: allow the materialization up to a maximum size of 1 million cells.
}

Or a named vector with at least one of
\itemize{
\item \code{cells} (numeric)
\item \code{rows} (numeric)
}

to allow materialization for data up to a certain size,
measured in cells (values) and rows in the resulting data frame.
The equivalent of \code{"drip"} is \code{c(cells = 1e6)}.

If \code{cells} is specified but not \code{rows}, \code{rows} is \code{Inf}.
If \code{rows} is specified but not \code{cells}, \code{cells} is \code{Inf}.

The default is to inherit the funneling of the input.}

\item{options}{A list of additional options to pass to create the storage format,
see \url{https://duckdb.org/docs/data/parquet/overview#writing-to-parquet-files}
or \url{https://duckdb.org/docs/data/csv/overview#writing-using-the-copy-statement}
for details.}
}
\description{
These functions apply to duckplyr frames.
They executes a query and stores the results in a flat file.
The result is a duckplyr frame that can be used with subsequent dplyr verbs.

\code{compute_parquet()} creates a Parquet file.

\code{compute_csv()} creates a CSV file.
}
\examples{
library(duckplyr)
df <- data.frame(x = c(1, 2))
df <- mutate(df, y = 2)
path <- tempfile(fileext = ".parquet")
df <- compute_parquet(df, path)
explain(df)
}
\seealso{
\code{\link[=compute.duckplyr_df]{compute.duckplyr_df()}}, \code{\link[dplyr:compute]{dplyr::collect()}}
}
