% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configuration options}
\description{
The behavior of duckplyr can be fine-tuned with several environment variables,
and one option.
}
\section{Environment variables}{


\code{DUCKPLYR_TEMP_DIR}: Set to a path where temporary files can be created.
By default, \code{\link[=tempdir]{tempdir()}} is used.

\code{DUCKPLYR_OUTPUT_ORDER}: If \code{TRUE}, row output order is preserved.
The default may change the row order where dplyr would keep it stable.
Preserving the order leads to more complicated execution plans
with less potential for optimization, and thus may be slower.

\code{DUCKPLYR_FORCE}: If \code{TRUE}, fail if duckdb cannot handle a request.

\code{DUCKPLYR_CHECK_ROUNDTRIP}: If \code{TRUE}, check if all columns are roundtripped perfectly
when creating a relational object from a data frame,
This is slow, and mostly useful for debugging.
The default is to check roundtrip of attributes.

\code{DUCKPLYR_METHODS_OVERWRITE}: If \code{TRUE}, call \code{methods_overwrite()}
when the package is loaded.

See \link{fallback} for more options related to printing, logging, and uploading
of fallback events.
}

