% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configuration options}
\description{
The behavior of duckplyr can be fine-tuned with several environment variables,
and one option.
}
\section{Options}{


\code{duckdb.materialize_message}: Set to \code{FALSE} to turn off diagnostic output from duckdb
on data frame materialization.
Currenty set to \code{TRUE} when duckplyr is loaded.
}

\section{Environment variables}{


\code{DUCKPLYR_OUTPUT_ORDER}: If \code{TRUE}, row output order is preserved.
If \code{FALSE},

\code{DUCKPLYR_FORCE}: If \code{TRUE}, fail if duckdb cannot handle a request.

\code{DUCKPLYR_FALLBACK_INFO}: If \code{TRUE}, print a message when a fallback to dplyr occurs
because duckdb cannot handle a request.

\code{DUCKPLYR_CHECK_ROUNDTRIP}: If \code{TRUE}, check if all columns are roundtripped perfectly
when creating a relational object from a data frame,
This is slow, and mostly useful for debugging.
The default is to check roundtrip of attributes.

\code{DUCKPLYR_EXPERIMENTAL}: If \code{TRUE}, pass \code{experimental = TRUE}
to certain duckdb functions.
Currently unused.
}

