% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckplyr-package.R
\docType{package}
\name{duckplyr-package}
\alias{duckplyr}
\alias{duckplyr-package}
\title{duckplyr: A DuckDB-backed version of dplyr}
\description{
Implements a subset of dplyr using DuckDB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/duckdblabs/duckplyr}
  \item Report bugs at \url{https://github.com/duckdblabs/duckplyr/issues}
}

}
\author{
\strong{Maintainer}: Hannes Mühleisen \email{hannes@duckdblabs.com} (\href{https://orcid.org/0000-0001-8552-0029}{ORCID})

Authors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
}

Other contributors:
\itemize{
  \item Posit [copyright holder, funder]
}

}
\keyword{internal}
