% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckplyr_execute.R
\name{duckplyr_execute}
\alias{duckplyr_execute}
\title{Execute a statement for the default connection}
\usage{
duckplyr_execute(sql)
}
\arguments{
\item{sql}{The statement to run.}
}
\value{
The return value of the \code{\link[DBI:dbExecute]{DBI::dbExecute()}} call, invisibly.
}
\description{
The \pkg{duckplyr} package relies on a DBI connection
to an in-memory database.
The \code{duckplyr_execute()} function allows running SQL statements
with this connection to, e.g., set up credentials
or attach other databases.
See \url{https://duckdb.org/docs/configuration/overview.html}
for more information on the configuration options.
}
\examples{
duckplyr_execute("SET threads TO 2")
}
