% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ducktbl.R
\name{ducktbl}
\alias{ducktbl}
\title{duckplyr data frames}
\usage{
ducktbl(...)
}
\arguments{
\item{...}{Passed on to \code{\link[=tibble]{tibble()}}.}
}
\value{
An object with the following classes:
\itemize{
\item \code{"duckplyr_df"}
\item Classes of a \link{tibble}
}
}
\description{
Data frames backed by duckplyr have a special class, \code{"duckplyr_df"},
in addition to the default classes.
This ensures that dplyr methods are dispatched correctly.
For such objects,
dplyr verbs such as \code{\link[=mutate]{mutate()}}, \code{\link[=select]{select()}} or \code{\link[=filter]{filter()}}  will attempt to use DuckDB.
If this is not possible, the original dplyr implementation is used.

\code{ducktbl()} works like \code{\link[=tibble]{tibble()}}.
In contrast to dbplyr, duckplyr data frames are "eager" by default.
To avoid unwanted expensive computation, they can be converted to "lazy" duckplyr frames
on which \code{\link[=collect]{collect()}} needs to be called explicitly.
}
\details{
Set the \code{DUCKPLYR_FALLBACK_INFO} and \code{DUCKPLYR_FORCE} environment variables
for more control over the behavior, see \link{config} for more details.
}
\examples{
x <- ducktbl(a = 1)
x

library(dplyr)
x \%>\%
  mutate(b = 2)

x$a

y <- ducktbl(a = 1, .lazy = TRUE)
y
try(y$a)
collect(y)$a
}
