% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain-rd.R, R/explain.R
\name{explain.duckplyr_df}
\alias{explain.duckplyr_df}
\title{Explain details of a tbl}
\usage{
\method{explain}{duckplyr_df}(x, ...)
}
\arguments{
\item{x}{An object to explain}

\item{...}{Other parameters possibly used by generic}
}
\description{
This is a method for the \code{\link[dplyr:explain]{dplyr::explain()}} generic.
This is a generic function which gives more details about an object
than \code{print()}, and is more focused on human readable output than \code{str()}.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true") || (rlang::is_interactive() && rlang::is_installed("dbplyr") && rlang::is_installed("Lahman"))) withAutoprint(\{ # examplesIf}
lahman_s <- dbplyr::lahman_sqlite()
batting <- tbl(lahman_s, "Batting")
explain(batting)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[dplyr:explain]{dplyr::explain()}}
}
