% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-expr.R
\name{expr}
\alias{expr}
\alias{new_relexpr}
\alias{relexpr_reference}
\alias{relexpr_constant}
\alias{relexpr_function}
\alias{relexpr_window}
\alias{relexpr_set_alias}
\title{Relational expressions}
\usage{
new_relexpr(x, class = NULL)

relexpr_reference(name, rel = NULL, alias = NULL)

relexpr_constant(val, alias = NULL)

relexpr_function(name, args, alias = NULL)

relexpr_window(
  expr,
  partitions,
  order_bys = list(),
  offset_expr = NULL,
  default_expr = NULL,
  alias = NULL
)

relexpr_set_alias(expr, alias = NULL)
}
\arguments{
\item{x}{An object.}

\item{class}{Classes added in front of the `"relational_relexpr"` base class.}

\item{name}{The name of the column or function to reference.}

\item{rel}{The name of the relation to reference.}

\item{alias}{An alias for the new expression.}

\item{val}{The value to use in the constant expression.}

\item{args}{Function arguments, a list of `expr` objects.}

\item{partitions}{Partitions, a list of `expr` objects.}

\item{order_bys}{which variables to order results by (list).}

\item{offset_expr}{offset relational expression.}

\item{default_expr}{default relational expression.}
}
\description{
These functions provide a backend-agnostic way to construct expression trees
built of column references, constants, and functions.
All subexpressions in an expression tree can have an alias.

`new_relexpr()` constructs an object of class `"relational_relexpr"`.
It is used by the higher-level constructors,
users should rarely need to call it directly.

`relexpr_reference()` constructs a reference to a column.

`relexpr_constant()` wraps a constant value.

`relexpr_function()` applies a function.
The arguments to this function are a list of other expression objects.

`relexpr_window()` marks an expression as a window operation,
with user-defined partitioning, order, offset, and default expressions.

`relexpr_set_alias()` updates the alias of a relational expression.
}
\examples{
relexpr_set_alias(alias = "my_predicate",
  relexpr_function("<",
    list(
      relexpr_window(
        relexpr_function("first", list(relexpr_reference("my_number"))),
        partitions = list(relexpr_reference("my_grouping"))
      ),
      relexpr_constant(42)
    )
  )
)
}
