% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-expr.R
\name{expr}
\alias{expr}
\alias{new_relexpr}
\alias{relexpr_reference}
\alias{relexpr_constant}
\alias{relexpr_function}
\alias{relexpr_window}
\alias{relexpr_set_alias}
\title{Relational expressions}
\usage{
new_relexpr(x, class = NULL)

relexpr_reference(name, rel = NULL, alias = NULL)

relexpr_constant(val, alias = NULL)

relexpr_function(name, args, alias = NULL)

relexpr_window(
  expr,
  partitions,
  order_bys = list(),
  offset_expr = NULL,
  default_expr = NULL,
  alias = NULL
)

relexpr_set_alias(expr, alias = NULL)
}
\arguments{
\item{x}{An object.}

\item{class}{Classes added in front of the \code{"relational_relexpr"} base class.}

\item{name}{The name of the column or function to reference.}

\item{rel}{The name of the relation to reference.}

\item{alias}{An alias for the new expression.}

\item{val}{The value to use in the constant expression.}

\item{args}{Function arguments, a list of \code{expr} objects.}

\item{expr}{An \code{expr} object.}

\item{partitions}{Partitions, a list of \code{expr} objects.}

\item{order_bys}{which variables to order results by (list).}

\item{offset_expr}{offset relational expression.}

\item{default_expr}{default relational expression.}
}
\description{
TBD.
}
