% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback.R
\name{fallback}
\alias{fallback}
\alias{fallback_sitrep}
\alias{fallback_review}
\alias{fallback_upload}
\alias{fallback_purge}
\alias{fallback_logging_optout}
\alias{fallback_logging_optin}
\alias{fallback_reporting_optin}
\alias{fallback_reporting_optout}
\alias{fallback_verbose_optin}
\alias{fallback_verbose_optout}
\title{Fallback to dplyr}
\usage{
fallback_sitrep()

fallback_review(oldest = NULL, newest = NULL, detail = TRUE)

fallback_upload(oldest = NULL, newest = NULL, strict = TRUE)

fallback_purge(oldest = NULL, newest = NULL)

fallback_logging_optout(purge = TRUE)

fallback_logging_optin()

fallback_reporting_optin()

fallback_reporting_optout()

fallback_verbose_optin()

fallback_verbose_optout()
}
\arguments{
\item{oldest, newest}{The number of oldest or newest reports to review.
If not specified, all reports are dispayed.}

\item{detail}{Print the full content of the reports.
Set to \code{FALSE} to only print the file names.}

\item{strict}{If \code{TRUE}, the function aborts if any of the reports fail to upload.
With \code{FALSE}, only a message is printed.}

\item{purge}{Logical. Whether to delete currently available reports.}
}
\description{
The \pkg{duckplyr} package aims at providing
a fully compatible drop-in replacement for \pkg{dplyr}.
To achieve this, only a carefully selected subset of dplyr's operations,
R functions, and R data types are implemented.
Whenever duckplyr encounters an incompatibility, it falls back to dplyr.

To assist future development, the fallback situations can be logged
to the console or to a local file and uploaded for analysis.
By default, \pkg{duckplyr} will not log or upload anything.
The functions and environment variables on this page control the process.

\code{fallback_sitrep()} prints the current settings for fallback logging and uploading,
the number of reports ready for upload, and the location of the logs.

\code{fallback_review()} prints the available reports for review to the console.

\code{fallback_upload()} uploads the available reports to a central server for analysis.
The server is hosted on AWS and the reports are stored in a private S3 bucket.
Only authorized personnel have access to the reports.

\code{fallback_purge()} deletes some or all available reports.

\code{fallback_logging_optout()} opts the machine out of fallback logging.

\code{fallback_logging_optin()} opts the machine in fallback logging.

\code{fallback_reporting_optin()} opts the machine in automatic report uploads.

\code{fallback_reporting_optout()} opts out the machine of automatic report uploads.

\code{fallback_verbose_optin()} opts the machine in messages about fallbacks.

\code{fallback_verbose_optout()} opts out the machine of messages about fallbacks.
}
\details{
Logging is opt-out but uploading is opt-in.

The following functions control the logging and uploading:
\itemize{
\item To opt the current machine out of logging, run \code{\link[=fallback_logging_optout]{fallback_logging_optout()}}.
\item To opt in again for logging, run \code{\link[=fallback_logging_optin]{fallback_logging_optin()}}.
\item \code{DUCKPLYR_FALLBACK_VERBOSE} controls printing, set it
to \code{TRUE} or \code{FALSE} to enable or disable printing.
If the value is \code{TRUE}, a message is printed to the console
for each fallback situation.
This setting is only relevant if logging is enabled.
\item To opt the current machine in automatic report uploads, run \code{\link[=fallback_reporting_optin]{fallback_reporting_optin()}}.
\item To opt out again of automatic report uploads, run \code{\link[=fallback_reporting_optout]{fallback_reporting_optout()}}.
\item \code{DUCKPLYR_FALLBACK_LOG_DIR} controls the location of the logs.
It must point to a directory (existing or not) where the logs will be written.
By default, logs are written to a directory in the user's cache directory
as returned by \code{tools::R_user_dir("duckplyr", "cache")}.
}

All code related to fallback logging and uploading is in the
\href{https://github.com/tidyverse/duckplyr/blob/main/R/fallback.R}{\code{fallback.R}} and
\href{https://github.com/tidyverse/duckplyr/blob/main/R/telemetry.R}{\code{telemetry.R}} files.
}
\examples{
fallback_sitrep()
}
