% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback.R
\name{fallback}
\alias{fallback}
\title{Fallback to dplyr}
\description{
The \pkg{duckplyr} package aims at providing
a fully compatible drop-in replacement for \pkg{dplyr}.
To achieve this, only a carefully selected subset of dplyr's operations,
R functions, and R data types are implemented.
Whenever duckplyr encounters an incompatibility, it falls back to dplyr.

To assist future development, the fallback situations can be logged
to a local file and uploaded for analysis.
The functions and environment variables on this page control the process.
}
\details{
Logging and uploading are both opt-in.
By default, for logging, a message is printed to the console
for the first time in a session and then once every 8 hours.
The number of available logs and instructions for uploading
are printed when the package is loaded.
}
