% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights.R
\name{flights_df}
\alias{flights_df}
\title{Flight data}
\usage{
flights_df()
}
\description{
Provides a variant of \code{nycflights13::flights} that is compatible with duckplyr,
as a tibble:
the timezone has been set to UTC to work around a current limitation of duckplyr, see \code{vignette("limits")}.
Call \code{\link[=as_duckdb_tibble]{as_duckdb_tibble()}} to enable duckplyr operations.
}
\examples{
\dontshow{if (requireNamespace("nycflights13", quietly = TRUE)) withAutoprint(\{ # examplesIf}
flights_df()
\dontshow{\}) # examplesIf}
}
