% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_duckplyr_df.R
\name{is_duckplyr_df}
\alias{is_duckplyr_df}
\title{Class predicate for duckplyr data frames}
\usage{
is_duckplyr_df(.data)
}
\arguments{
\item{.data}{The object to test}
}
\value{
\code{TRUE} if the input object is of class \code{"duckplyr_df"},
otherwise \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Tests if the input object is of class \code{"duckplyr_df"}.
}
\examples{
tibble(a = 1:3) |>
  is_duckplyr_df()

tibble(a = 1:3) |>
  as_duckplyr_df() |>
  is_duckplyr_df()
}
\keyword{internal}
