% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{data}
\name{methods_overwrite}
\alias{methods_overwrite}
\alias{methods_restore}
\title{Forward all dplyr methods to duckplyr}
\format{
An object of class \code{function} of length 1.
}
\usage{
methods_overwrite
}
\value{
Called for their side effects.
}
\description{
After calling \code{methods_overwrite()}, all dplyr methods are redirected to duckplyr
for the duraton of the session, or until a call to \code{methods_restore()}.
}
\keyword{datasets}
