% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{data}
\name{methods_overwrite}
\alias{methods_overwrite}
\alias{methods_restore}
\title{Forward all dplyr methods to duckplyr}
\format{
An object of class \code{function} of length 1.
}
\usage{
methods_overwrite
}
\description{
Calling this function
}
\keyword{datasets}
