% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{new_relational}
\alias{new_relational}
\title{Relational API}
\usage{
new_relational(..., class = NULL)
}
\arguments{
\item{...}{Passed on to [structure()]}

\item{class}{Classes added in front of the `"relational"` base class}
}
\value{
A (new/modified) relational object.
}
\description{
TBD.
}
