% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckplyr-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{tibble}
\alias{across}
\alias{add_count}
\alias{add_tally}
\alias{all_equal}
\alias{anti_join}
\alias{arrange}
\alias{auto_copy}
\alias{c_across}
\alias{case_when}
\alias{collapse}
\alias{collect}
\alias{compute}
\alias{count}
\alias{cross_join}
\alias{desc}
\alias{distinct}
\alias{explain}
\alias{filter}
\alias{first}
\alias{full_join}
\alias{if_all}
\alias{if_any}
\alias{inner_join}
\alias{intersect}
\alias{join_by}
\alias{lag}
\alias{last}
\alias{lead}
\alias{left_join}
\alias{mutate}
\alias{n}
\alias{n_distinct}
\alias{nest_by}
\alias{nest_join}
\alias{nth}
\alias{pull}
\alias{reframe}
\alias{relocate}
\alias{rename}
\alias{rename_with}
\alias{right_join}
\alias{row_number}
\alias{rows_append}
\alias{rows_delete}
\alias{rows_insert}
\alias{rows_patch}
\alias{rows_update}
\alias{rows_upsert}
\alias{sample_frac}
\alias{sample_n}
\alias{select}
\alias{semi_join}
\alias{setdiff}
\alias{setequal}
\alias{slice}
\alias{slice_head}
\alias{slice_max}
\alias{slice_min}
\alias{slice_sample}
\alias{slice_tail}
\alias{summarise}
\alias{summarize}
\alias{symdiff}
\alias{tally}
\alias{tbl_vars}
\alias{transmute}
\alias{ungroup}
\alias{union}
\alias{union_all}
\alias{with_groups}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{across}}, \code{\link[dplyr:count]{add_count}}, \code{\link[dplyr:count]{add_tally}}, \code{\link[dplyr]{all_equal}}, \code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{auto_copy}}, \code{\link[dplyr]{c_across}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr:compute]{collapse}}, \code{\link[dplyr:compute]{collect}}, \code{\link[dplyr]{compute}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{cross_join}}, \code{\link[dplyr]{desc}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{explain}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:nth]{first}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr:across]{if_all}}, \code{\link[dplyr:across]{if_any}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:setops]{intersect}}, \code{\link[dplyr]{join_by}}, \code{\link[dplyr:lead-lag]{lag}}, \code{\link[dplyr:nth]{last}}, \code{\link[dplyr:lead-lag]{lead}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:context]{n}}, \code{\link[dplyr]{n_distinct}}, \code{\link[dplyr]{nest_by}}, \code{\link[dplyr]{nest_join}}, \code{\link[dplyr]{nth}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{reframe}}, \code{\link[dplyr]{relocate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:rename]{rename_with}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr]{row_number}}, \code{\link[dplyr:rows]{rows_append}}, \code{\link[dplyr:rows]{rows_delete}}, \code{\link[dplyr:rows]{rows_insert}}, \code{\link[dplyr:rows]{rows_patch}}, \code{\link[dplyr:rows]{rows_update}}, \code{\link[dplyr:rows]{rows_upsert}}, \code{\link[dplyr:sample_n]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr:setops]{setdiff}}, \code{\link[dplyr:setops]{setequal}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_head}}, \code{\link[dplyr:slice]{slice_max}}, \code{\link[dplyr:slice]{slice_min}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:slice]{slice_tail}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr:setops]{symdiff}}, \code{\link[dplyr:count]{tally}}, \code{\link[dplyr]{tbl_vars}}, \code{\link[dplyr:reexports]{tibble}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}, \code{\link[dplyr:setops]{union}}, \code{\link[dplyr:setops]{union_all}}, \code{\link[dplyr]{with_groups}}}
}}

