% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_alias}
\alias{rel_alias}
\title{Get the internal alias for a relation object}
\usage{
rel_alias(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
rel_alias(rel)
}
