% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_distinct}
\alias{rel_distinct}
\title{Lazily compute a distinct result on a relation object}
\usage{
rel_distinct(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object with distinct rows
}
\description{
TBD.
}
\examples{
rel <- new_relational(c("a", "a", "b"))
rel_distinct.relational <- function(rel, ...) new_relational(unique(df))
rel_distinct(rel)
}
