% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_distinct}
\alias{rel_distinct}
\title{Lazily compute a distinct result on a relation object}
\usage{
rel_distinct(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object with distinct rows
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
rel2 <- rel_distinct(rel)
}
