% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_filter}
\alias{rel_filter}
\title{Lazily filter a relation object}
\usage{
rel_filter(rel, exprs, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{exprs}{a list of DuckDB expressions to filter by}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
the now filtered relation object
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
rel2 <- rel_filter(rel, list(expr_function("gt", list(expr_reference("cyl"), expr_constant("6")))))
}
