% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_join}
\alias{rel_join}
\title{Lazily INNER join two relational objects}
\usage{
rel_join(left, right, conds, join, ...)
}
\arguments{
\item{left}{the left-hand-side relational object}

\item{right}{the right-hand-side relational object}

\item{conds}{a list of DuckDB expressions to use for the join}

\item{join}{type of join}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relational object resulting from the join
}
\description{
TBD.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
left <- rel_from_df(mtcars)
right <- rel_from_df(mtcars)
cond <- list(
  relexpr_function(
    "eq",
    list(relexpr_reference("cyl", left), relexpr_reference("cyl", right))
  )
)
rel2 <- rel_join(left, right, cond)
\dontshow{\}) # examplesIf}
}
