% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_limit}
\alias{rel_limit}
\title{Lazily limit the rows in a relation object}
\usage{
rel_limit(rel, n, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{n}{The number of rows.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A (new/modified) relational object.
}
\description{
TBD.
}
