% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_set_alias}
\alias{rel_set_alias}
\title{Set the internal alias for a relation object}
\usage{
rel_set_alias(rel, alias, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{alias}{the new alias}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A (new/modified) relational object.
}
\description{
TBD.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rel <- rel_from_df(mtcars)
rel_set_alias(rel, "my_new_alias")
\dontshow{\}) # examplesIf}
}
