% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.R
\name{rel_to_df}
\alias{rel_to_df}
\title{Convert a relation object to a data frame}
\usage{
rel_to_df(rel, ...)
}
\arguments{
\item{rel}{The relation object.}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
A data frame.
}
\description{
TBD.
}
\examples{
rel <- rel_from_df(mtcars)
rel2 <- rel_filter(rel, list(expr_function("gt", list(expr_reference("cyl"), expr_constant("6")))))
}
