% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_union_all}
\alias{rel_union_all}
\title{Lazily compute a set_union_all result on a relational object}
\usage{
rel_union_all(rel_a, rel_b, ...)
}
\arguments{
\item{rel_a}{a DuckDB relational object}

\item{rel_b}{a DuckDB relational object}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relational object with the result
}
\description{
TBD.
}
\examples{
rel_a <- new_relational(c(1, 1, 2), class = "vecrel")
rel_b <- new_relational(c(1, 3, 2), class = "vecrel")
rel_union_all.vecrel <- function(rel_a, rel_b, ...) {
  new_relational(union(rel_a, rel_b), class = class(rel_a))
}
rel_union_all(rel_a, rel_b)
}
