% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational-rel.R
\name{rel_union_all}
\alias{rel_union_all}
\title{Lazily compute a set_union_all result on a relation object}
\usage{
rel_union_all(rel_a, rel_b, ...)
}
\arguments{
\item{rel_a}{a DuckDB relation object}

\item{rel_b}{a DuckDB relation object}

\item{...}{Reserved for future extensions, must be empty.}
}
\value{
a new relation object with the result
}
\description{
TBD.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rel <- rel_from_df(mtcars)
rel2 <- rel_union_all(rel)
\dontshow{\}) # examplesIf}
}
